+++ packages/ui/components/InformationCard/InformationCardToolTip.tsx
@@ -10,35 +10,81 @@ export interface InformationCardTooltipProps
     delayDuration?: number;
   }
 
+const InformationCardToolTipTrigger = ({ className, ...props }) => {
+  return (
+    <span
+      className={cn(
+        "bg-orange rounded-full cursor-pointer text-md text-white ml-2 font-proxima p-1 leading-4 w-4 h-4 inline-flex items-center justify-center",
+        className,
+      )}
+      {...props}
+    >
+      ?
+    </span>
+  );
+}
+
 const InformationCardToolTip = React.forwardRef<
   HTMLDivElement,
   InformationCardTooltipProps
 >(({ className, children, delayDuration = 150, ...props }, ref) => {
+  const innerRef = React.useRef<HTMLDivElement | null>(null);
+
+  React.useImperativeHandle(ref, () => innerRef.current!, []);
+
+  const [isOpen, setIsOpen] = React.useState(props.defaultOpen ?? false);
+  const [isTouchDevice, setIsTouchDevice] = React.useState(false);
+
+  const toggleMobileOpen = (e) => {
+    e.preventDefault();
+    setIsOpen(true);
+  }
+
+  React.useEffect(() => {
+    const updateScreen = () => {
+      // Update the state to reflect the current screen size
+      // 992px is the breakpoint for lg (tailwind config)
+      setIsTouchDevice(window?.document.body.offsetWidth < 992);
+    }
+    updateScreen();
+    window.addEventListener('resize', updateScreen)
+		return () => window.removeEventListener('resize', updateScreen)
+  }, [ref]);
+
+  React.useEffect(() => {
+    const handleClickOutside = (event) => {
+      if (innerRef?.current && !innerRef?.current.contains(event.target)) {
+        setIsOpen(false);
+      }
+    }
+    document.addEventListener("mousedown", handleClickOutside);
+    return () => {
+      document.removeEventListener("mousedown", handleClickOutside);
+    };
+  }, [ref]);
+
+
   return (
     <TooltipProvider>
-    <Tooltip delayDuration={delayDuration} {...props}>
-      <TooltipTrigger>
-        
-          <span
-            className={cn(
-              "bg-orange rounded-full cursor-pointer text-md text-white ml-2 font-proxima p-1 leading-4 w-4 h-4 inline-flex items-center justify-center",
-              className,
-            )}
-          >
-            ?
-          </span>
+      <Tooltip 
+        delayDuration={!isTouchDevice ? delayDuration : 0}
+        {...(isTouchDevice ? { open: isOpen } : {} )}
+        {...props} 
+      >
+        <TooltipTrigger onClick={isTouchDevice ? toggleMobileOpen : () => {}}>
+          <InformationCardToolTipTrigger className={className} />
         </TooltipTrigger>
         <TooltipContent
           className={cn(
             "bg-white rounded-md shadow-lg p-4 max-w-80 font-normal text-lg text-darkblue font-proxima leading-5 border-neutral-400 border text-wrap",
             className,
           )}
-          ref={ref}
+          ref={innerRef}
         >
           {children}
         </TooltipContent>
       </Tooltip>
-      </TooltipProvider>
+    </TooltipProvider>
   );
 });